+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- fixed alias resolving when using as argument to `generate`
+
 ## [1.4.0] - 2023-05-12
 ### Changed
 - moved default config to separate file `.devcontainer_manager/templates/default.yaml`